refactor: Rename realtimes.js to realtime.js
